rename schemas folder to 'models'
